Add newsletter form api
